requirements.txt
annotated-types==0.6.0
anyio==3.7.1
APScheduler==3.10.4
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.4.2
docker==6.1.3
exceptiongroup==1.2.0
fastapi==0.104.1
h11==0.14.0
idna==3.6
kafka-python==2.0.2
numpy==1.26.2
packaging==23.2
pandas==2.1.3
pydantic==2.5.2
pydantic_core==2.14.5
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
uvicorn==0.24.0.post1
websocket-client==1.7.0
